mixin multi pdf renderers
